Add frontend boilerplate and basic functionalities.
I've added some basic code for the front end that was loosely based off the contacts app v5 design.

Still early prototype with not much working as of right now. Some code is just temporary placeholder till we find something better.

I've also added a leaflet.js map to the homepage. The website felt empty and  I enjoy having something there while working on it. Not sure if this is the best option or not. If we decide not to use we can remove it easily.

Added some sample jquery and an ajax calls to our server in frontend.js.

It's late and I've probably made mistakes so please reach out if something not right.