Correct queries arrive at front end
Sorry for the stupid large PR. I think I have most of what we discussed worked out. When I select the filters now and do the search the correct results are displayed in the browser console for both usages and responses!

for Usage.js and Response.js I added some code to handle the arrays of geo and the age edge case with Response. The mongo queries seem to be working as intended with multiple geos.

for Usages.js and Responses.js I did some work on the validators and they seem to be working alright.

frontend.js had a bunch or small changes. my geo validator was buggy and I sorted the issue with getResponses and getUsages not returning anything.

There's a known issue with the map not painting correctly when a input box starts checked that I'll fix tomorrow.